pyproject.toml
[tool.poetry]
name = "chat2vis-streamlit"
version = "0.1.0"
description = "An ephemeral graph generated through a user interaction"
authors = ["snyata <core@synavate.tech>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
uvicorn = "^0.23.1"
python = ">=3.10,<3.12"
cogdb = "^3.0.7"
python-dotenv = "^1.0.1"
openai = "^1.12.0"
streamlit = "^1.31.1"
pandas = "^2.2.0"
langchain-openai = "^0.0.6"
tiktoken = "^0.6.0"
langchain = "^0.1.7"
scikit-learn = "^1.4.1"
numpy = "^1.26.4"
qdrant-client = "^1.7.3"
loguru = "^0.7.2"
motor = "^3.3.2"
sqlalchemy = "^2.0.28"
tqdm = "^4.66.2"
llama-index-core = "^0.10.20.post2"
llama-index-llms-openai = "^0.1.12"
llama-index-llms-replicate = "^0.1.3"
llama-index-embeddings-huggingface = "^0.1.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
